Faire de la watcHApiRegistre une classe interne
